broke out api testing boilerplate into traits and interfaces, for easier mixin/matching.
