+++ tests/IliosApiBundle/DeleteEndpointTestInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Tests\IliosApiBundle;
+
+/**
+ * Interface DeleteEndpointTestInterface
+ * @package Tests\IliosApiBundle
+ */
+interface DeleteEndpointTestInterface
+{
+    /**
+     * Test deleting data
+     */
+    public function testDelete();
+}
